lambda/requirements.txt
PyJWT==2.8.0
boto3==1.34.34
